services/audio_bridge/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
pydantic==2.8.2
httpx==0.27.0
pyaudio==0.2.11
numpy==1.24.3
python-dotenv==1.0.0
